Add files via upload
This is not my work, just a modified version that works for my setup. The naming convention in the file was used from xivbar, which if you had xivbar as well, when the call for additional local files was done, it would recursively look through the folder structure, so it pulled from the wrong folder. Adjusted the data inside to pull from the correct folder structure.